Add the basic actions to manage vehicles
Also, properly map the relations between the vehicles and the
characteristics and add the reverse mapping between category and
characteristics to be able to group characteristics by category
and nicely display them on the form for vehicles.

Remove some useless stuff that was generated by Eclipse.

Regarding the collection loading related stuff, by default, JPA
considers that all association are lazy loaded. We are either
forcing this loading by using an "eager" strategy when it won't
produce a performance bottleneck or trying to load the collection
while we are still in a transactional context. The latter seems
very hacky but works so ¯\_(ツ)_/¯