Fix the displaying of deliveries
Since we are eagerly loading the possible characteristics of a
vehicle and characteristics are also attached to a purchase, JPA
was complaining that about multi bag loading if we were to load
both of these collections at the same moment.

Thus, we just need to explicitly load the collection of characteristics
on either of these collections. Since we are only using deliveries
in a single place, let's make the call explicit there.